+++ Lucc/Frontend/AST.h
@@ -4,6 +4,7 @@
 #include <string>
 #include "Type.h"
 #include "Symbol.h"
+#include "SourceLocation.h"
 
 namespace lucc
 {
@@ -49,39 +50,39 @@ namespace lucc
 	{
 	public:
 		virtual ~INodeVisitorAST() = default;
-		virtual void Visit(NodeAST const& node, size_t depth) {}
-		virtual void Visit(TranslationUnitAST const& node, size_t depth) {}
-		virtual void Visit(ExprAST const& node, size_t depth) {}
-		virtual void Visit(UnaryExprAST const& node, size_t depth) {}
-		virtual void Visit(BinaryExprAST const& node, size_t depth) {}
-		virtual void Visit(TernaryExprAST const& node, size_t depth) {}
-		virtual void Visit(FunctionCallExprAST const& node, size_t depth) {}
-		virtual void Visit(CastExprAST const& node, size_t depth) {}
-		virtual void Visit(IntLiteralAST const& node, size_t depth) {}
-		virtual void Visit(StringLiteralAST const& node, size_t depth) {}
-		virtual void Visit(IdentifierExprAST const& node, size_t depth) {}
-		virtual void Visit(DeclRefExprAST const& node, size_t depth) {}
-		virtual void Visit(MemberRefExprAST const& node, size_t depth) {}
-		virtual void Visit(StmtAST const& node, size_t depth) {}
-		virtual void Visit(CompoundStmtAST const& node, size_t depth) {}
-		virtual void Visit(DeclStmtAST const& node, size_t depth) {}
-		virtual void Visit(ExprStmtAST const& node, size_t depth) {}
-		virtual void Visit(NullStmtAST const& node, size_t depth) {}
-		virtual void Visit(IfStmtAST const& node, size_t depth) {}
-		virtual void Visit(WhileStmtAST const& node, size_t depth) {}
-		virtual void Visit(DoWhileStmtAST const& node, size_t depth) {}
-		virtual void Visit(ForStmtAST const& node, size_t depth) {}
-		virtual void Visit(SwitchStmtAST const& node, size_t depth) {}
-		virtual void Visit(CaseStmtAST const& node, size_t depth) {}
-		virtual void Visit(ReturnStmtAST const& node, size_t depth) {}
-		virtual void Visit(GotoStmtAST const& node, size_t depth) {}
-		virtual void Visit(LabelStmtAST const& node, size_t depth) {}
-		virtual void Visit(BreakStmtAST const& node, size_t depth) {}
-		virtual void Visit(ContinueStmtAST const& node, size_t depth) {}
-		virtual void Visit(DeclAST const& node, size_t depth) {}
-		virtual void Visit(VarDeclAST const& node, size_t depth) {}
-		virtual void Visit(FunctionDeclAST const& node, size_t depth) {}
-		virtual void Visit(TypedefDeclAST const& node, size_t depth) {}
+		virtual void Visit(NodeAST const& node, uint32 depth) {}
+		virtual void Visit(TranslationUnitAST const& node, uint32 depth) {}
+		virtual void Visit(ExprAST const& node, uint32 depth) {}
+		virtual void Visit(UnaryExprAST const& node, uint32 depth) {}
+		virtual void Visit(BinaryExprAST const& node, uint32 depth) {}
+		virtual void Visit(TernaryExprAST const& node, uint32 depth) {}
+		virtual void Visit(FunctionCallExprAST const& node, uint32 depth) {}
+		virtual void Visit(CastExprAST const& node, uint32 depth) {}
+		virtual void Visit(IntLiteralAST const& node, uint32 depth) {}
+		virtual void Visit(StringLiteralAST const& node, uint32 depth) {}
+		virtual void Visit(IdentifierExprAST const& node, uint32 depth) {}
+		virtual void Visit(DeclRefExprAST const& node, uint32 depth) {}
+		virtual void Visit(MemberRefExprAST const& node, uint32 depth) {}
+		virtual void Visit(StmtAST const& node, uint32 depth) {}
+		virtual void Visit(CompoundStmtAST const& node, uint32 depth) {}
+		virtual void Visit(DeclStmtAST const& node, uint32 depth) {}
+		virtual void Visit(ExprStmtAST const& node, uint32 depth) {}
+		virtual void Visit(NullStmtAST const& node, uint32 depth) {}
+		virtual void Visit(IfStmtAST const& node, uint32 depth) {}
+		virtual void Visit(WhileStmtAST const& node, uint32 depth) {}
+		virtual void Visit(DoWhileStmtAST const& node, uint32 depth) {}
+		virtual void Visit(ForStmtAST const& node, uint32 depth) {}
+		virtual void Visit(SwitchStmtAST const& node, uint32 depth) {}
+		virtual void Visit(CaseStmtAST const& node, uint32 depth) {}
+		virtual void Visit(ReturnStmtAST const& node, uint32 depth) {}
+		virtual void Visit(GotoStmtAST const& node, uint32 depth) {}
+		virtual void Visit(LabelStmtAST const& node, uint32 depth) {}
+		virtual void Visit(BreakStmtAST const& node, uint32 depth) {}
+		virtual void Visit(ContinueStmtAST const& node, uint32 depth) {}
+		virtual void Visit(DeclAST const& node, uint32 depth) {}
+		virtual void Visit(VarDeclAST const& node, uint32 depth) {}
+		virtual void Visit(FunctionDeclAST const& node, uint32 depth) {}
+		virtual void Visit(TypedefDeclAST const& node, uint32 depth) {}
 	};
 
 	class x86_64Context;
@@ -91,7 +92,7 @@ namespace lucc
 	{
 	public:
 		virtual ~NodeAST() = default;
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const = 0;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const = 0;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const {}
 
 	protected:
@@ -107,7 +108,7 @@ namespace lucc
 		}
 		auto const& GetDeclarations() const { return declarations; }
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
@@ -134,7 +135,7 @@ namespace lucc
 		VarSymbol const& GetSymbol() const { return sym; }
 
 		virtual int32 GetLocalOffset() const { return 0; }
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 
 	protected:
 		std::string name;
@@ -159,7 +160,7 @@ namespace lucc
 		bool IsGlobal() const { return GetSymbol().global; }
 		virtual int32 GetLocalOffset() const override { return local_offset; }
 		ExprAST const* GetInitExpr() const { return init_expr.get(); }
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
@@ -199,7 +200,7 @@ namespace lucc
 			for (auto const* local : local_variables) fn(local);
 		}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
@@ -216,7 +217,7 @@ namespace lucc
 	{
 	public:
 		TypedefDeclAST(std::string_view typedef_name) : DeclAST(typedef_name, DeclKind::Typedef) {}
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 	};
 
 	enum class StmtKind
@@ -241,7 +242,7 @@ namespace lucc
 	{
 	public:
 		StmtKind GetStmtKind() const { return kind; }
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 
 	protected:
 		StmtKind kind;
@@ -260,7 +261,7 @@ namespace lucc
 			statements.push_back(std::move(stmt));
 		}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
@@ -272,7 +273,7 @@ namespace lucc
 		ExprStmtAST(std::unique_ptr<ExprAST>&& expr) : StmtAST(expr ? StmtKind::Expr : StmtKind::Null), expr(std::move(expr)) {}
 		ExprAST* GetExpr() const { return expr.get(); }
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
@@ -284,7 +285,7 @@ namespace lucc
 		DeclStmtAST(std::vector<std::unique_ptr<DeclAST>>&& decls) : StmtAST(StmtKind::Decl), decls(std::move(decls)) {}
 		auto const& GetDeclarations() const { return decls; }
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
@@ -294,7 +295,7 @@ namespace lucc
 	{
 	public:
 		NullStmtAST() : ExprStmtAST(nullptr) {}
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 	};
 	class IfStmtAST final : public StmtAST
 	{
@@ -314,7 +315,7 @@ namespace lucc
 			else_stmt = std::move(_else_stmt);
 		}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
@@ -328,7 +329,7 @@ namespace lucc
 		CaseStmtAST() : StmtAST(StmtKind::Case), switch_id(-1), is_default(true), value(0), label_name("L_default") {}
 		explicit CaseStmtAST(int64 value) : StmtAST(StmtKind::Case), switch_id(-1), is_default(false), value(value), label_name("L_case" + std::to_string(value)) {}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 		void SetSwitchId(uint64 _switch_id) { switch_id = _switch_id; }
@@ -346,7 +347,7 @@ namespace lucc
 	public:
 		explicit GotoStmtAST(std::string_view label) : StmtAST(StmtKind::Goto), goto_label(label) {}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 		std::string_view GetLabel() const { return goto_label; }
@@ -359,7 +360,7 @@ namespace lucc
 	public:
 		LabelStmtAST(std::string_view label) : StmtAST(StmtKind::Label), label_name(label) {}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 		std::string_view GetLabel() const { return label_name; }
@@ -372,7 +373,7 @@ namespace lucc
 	public:
 		BreakStmtAST() : StmtAST(StmtKind::Break), label_id(-1) {}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override
 		{
 			visitor.Visit(*this, depth);
 		}
@@ -389,7 +390,7 @@ namespace lucc
 	public:
 		ContinueStmtAST() : StmtAST(StmtKind::Continue), label_id(-1) {}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override
 		{
 			visitor.Visit(*this, depth);
 		}
@@ -406,7 +407,7 @@ namespace lucc
 	public:
 		SwitchStmtAST() : StmtAST(StmtKind::Switch) {}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 		void SetCondition(std::unique_ptr<ExprAST>&& _condition)
@@ -448,7 +449,7 @@ namespace lucc
 	public:
 		WhileStmtAST() : StmtAST(StmtKind::While) {}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 		void SetCondition(std::unique_ptr<ExprAST>&& _condition)
@@ -480,7 +481,7 @@ namespace lucc
 	public:
 		DoWhileStmtAST() : StmtAST(StmtKind::DoWhile) {}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 		void SetCondition(std::unique_ptr<ExprAST>&& _condition)
@@ -512,7 +513,7 @@ namespace lucc
 	public:
 		ForStmtAST() : StmtAST(StmtKind::For) {}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 		void SetInit(std::unique_ptr<StmtAST>&& _init)
@@ -555,7 +556,7 @@ namespace lucc
 		explicit ReturnStmtAST(std::unique_ptr<ExprStmtAST>&& ret_expr)
 			: StmtAST(StmtKind::Return), ret_expr(std::move(ret_expr)) {}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
@@ -603,7 +604,7 @@ namespace lucc
 	class ExprAST : public NodeAST
 	{
 	public:
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 
 		virtual bool IsConstexpr() const { return false; }
 		virtual int64 EvaluateConstexpr() const { return 0; }
@@ -645,7 +646,7 @@ namespace lucc
 		UnaryExprKind GetUnaryKind() const { return op; }
 		ExprAST* GetOperand() const { return operand.get(); }
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 		virtual bool IsConstexpr() const override;
@@ -671,7 +672,7 @@ namespace lucc
 		ExprAST* GetLHS() const { return lhs.get(); }
 		ExprAST* GetRHS() const { return rhs.get(); }
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 		virtual bool IsConstexpr() const override;
 		virtual int64 EvaluateConstexpr() const override;
@@ -696,7 +697,7 @@ namespace lucc
 		void SetTrueExpr(std::unique_ptr<ExprAST>&& expr) { true_expr = std::move(expr); }
 		void SetFalseExpr(std::unique_ptr<ExprAST>&& expr) { false_expr = std::move(expr); }
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 		virtual bool IsConstexpr() const override;
 		virtual int64 EvaluateConstexpr() const override;
@@ -737,7 +738,7 @@ namespace lucc
 		ExprAST* GetFunction() const { return func_expr.get(); }
 		std::vector<std::unique_ptr<ExprAST>> const& GetFunctionArgs() const { return func_args; }
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
@@ -750,7 +751,7 @@ namespace lucc
 		IntLiteralAST(int64 value, SourceLocation const& loc) : ExprAST(ExprKind::IntLiteral, loc, builtin_types::Int), value(value) {}
 		int64 GetValue() const { return value; }
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 		virtual bool IsConstexpr() const override;
 		virtual int64 EvaluateConstexpr() const override;
@@ -761,10 +762,10 @@ namespace lucc
 	class StringLiteralAST final : public ExprAST
 	{
 	public:
-		StringLiteralAST(std::string_view str, SourceLocation const& loc) : ExprAST(ExprKind::StringLiteral, loc, QualifiedType(ArrayType(builtin_types::Char, str.size()))), str(str) {}
+		StringLiteralAST(std::string_view str, SourceLocation const& loc) : ExprAST(ExprKind::StringLiteral, loc, QualifiedType(ArrayType(builtin_types::Char, (uint32)str.size()))), str(str) {}
 		std::string_view GetString() const { return str; }
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
@@ -785,7 +786,7 @@ namespace lucc
 			SetCastType();
 		}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
@@ -839,7 +840,7 @@ namespace lucc
 
 		DeclAST const* GetDeclaration() const { return decl_ast; }
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	protected:
@@ -863,7 +864,7 @@ namespace lucc
 			return decl_ast->GetLocalOffset() + member_offset;
 		}
 
-		virtual void Accept(INodeVisitorAST& visitor, size_t depth) const override;
+		virtual void Accept(INodeVisitorAST& visitor, uint32 depth) const override;
 		virtual void Codegen(x86_64Context& ctx, Register* result = nullptr) const override;
 
 	private:
